Fix reliance on undefined behaviour in GetNumProcessors() on Windows

+++ Code/Core/Env/Env.cpp
@@ -85,7 +85,7 @@
 
             for ( size_t processorID = 0; processorID < maxLogicalProcessorsInThisGroup; ++processorID )
             {
-                numProcessorsInThisGroup += ( ( groupProcessorMask.Mask & ( 1i64 << processorID ) ) != 0 ) ? 1 : 0;
+                numProcessorsInThisGroup += ( ( groupProcessorMask.Mask & ( uint64_t(1) << processorID ) ) != 0 ) ? 1 : 0;
             }
 
             numProcessorsInAllGroups += numProcessorsInThisGroup;
